Use double instead of float for transforms

+++ src/Transform.cpp
@@ -1,5 +1,5 @@
 #include "Transform.hpp"
 
-Transform::Transform(int type, int axis, float value) :
+Transform::Transform(int type, int axis, double value) :
     type(type), axis(axis), value(value) {
 }
